New improved ServiceManager + test case. Modules can now register services they offer to the service manager. Other modules can then use these services through service interfaces.
Fixed few warnings with Core::Exception.

git-svn-id: https://realxtend-naali.googlecode.com/svn/trunk@54 5b2332b8-efa3-11de-8684-7d64432d61a3

+++ Core/CoreException.h
@@ -21,7 +21,7 @@ namespace Core
         {
             do_free_ = 1;
 
-            int len = strlen(what);
+            size_t len = strlen(what);
             what_ = static_cast< char* >(malloc(len + 1));
             strcpy(const_cast< char* >(what_), what);
         }
@@ -33,7 +33,7 @@ namespace Core
 
             if (do_free_)
             {
-                int len = strlen(other.what_);
+                size_t len = strlen(other.what_);
                 what_ = static_cast< char* >(malloc(len + 1));
                 strcpy(const_cast< char* >(what_), other.what_);
             } else
@@ -51,7 +51,7 @@ namespace Core
 
                 if (do_free_)
                 {
-                    int len = strlen(other.what_);
+                    size_t len = strlen(other.what_);
                     what_ = static_cast< char* >(malloc(len + 1));
                     strcpy(const_cast< char* >(what_), other.what_);
                 } else
